restructuring code base
