+++ client/scard_v2.txt
