+++ old/analysis/timelogging/src/alllogs.py
