+++ old/analysis/timelogging/src/log_write.py
