+++ old/analysis/timelogging/src/timeparser.py
