+++ server/src/new_script_generator.py
