+++ server/src/script_generators/startup.py
