+++ server/src/utils/user_validation.py
